+++ cloudsplaining/scan/policy_document.py
@@ -1,5 +1,4 @@
 import logging
-from policy_sentry.querying.actions import get_action_data, remove_actions_not_matching_access_level
 from cloudsplaining.scan.statement_details import StatementDetails
 from cloudsplaining.shared.constants import PRIVILEGE_ESCALATION_METHODS
 from cloudsplaining.shared.constants import READ_ONLY_DATA_LEAK_ACTIONS
@@ -24,8 +23,6 @@ class PolicyDocument:
         for statement in statement_structure:
             self.statements.append(StatementDetails(statement))
 
-        self.check_statements_on_intake()
-
     @property
     def json(self):
         """Return the Policy in JSON"""
@@ -39,11 +36,6 @@ class PolicyDocument:
         allowed_actions = list(dict.fromkeys(allowed_actions))
         return allowed_actions
 
-    def check_statements_on_intake(self):
-        # Just doing this to get the error message if it exists. Probably a better way of doing this.
-        result = self.contains_statement_using_not_action
-        result = self.contains_not_resource_with_allow
-
     @property
     def contains_statement_using_not_action(self):
         """If NotAction is used, flag it so the assessor can triage manually"""
@@ -53,37 +45,9 @@ class PolicyDocument:
             if statement.not_action:
                 not_action_statements.append(statement.json)
                 if not statement.has_resource_constraints and statement.effect_allow:
-                    print(f"{RED}\tThe policy has Effect=Allow and uses NotAction without any resource "
+                    print(f"{RED}\tNOTE: The policy has Effect=Allow and uses NotAction without any resource "
                           f"constraints: {statement.json}{RESET}")
         return not_action_statements
-
-    @property
-    def contains_not_resource_with_allow(self):
-        """Per the AWS documentation, the NotResource should never be used with the Allow Effect.
-        See documentation here. https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_elements_notresource.html#notresource-element-combinations"""
-        # If any of these are ever true, "ring the bell"
-        not_resource_status = None
-        effect_allow_status = None
-        result = False
-        for statement in self.statements:
-            if statement.not_resource:
-                not_resource_status = True
-            if statement.effect_allow:
-                effect_allow_status = True
-        # If either of these are ever true. Per AWS docs:
-        #   'Be careful using the NotResource element and "Effect": "Allow" in the same statement
-        #   or in a different statement within a policy.'
-        if not_resource_status and effect_allow_status:
-            result = True
-            # TODO: Must have this as part of the report before open sourcing
-            print(f"{RED}\tThe policy has Effect=Allow and uses NotResource. This is DANGEROUS. NotResource allows"
-                  f" all services and resources that are not explicitly listed, and could result in granting "
-                  f"users more permissions than you intended. AWS Documentation specifically says that you should NOT"
-                  f" use this. The policy is as follows: {self.policy}"
-                  f""
-                  f"{RESET}")
-
-        return result
 
     @property
     def allows_privilege_escalation(self):