+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # CHANGELOG
+
+## Unreleased
+* Removed "Stats for Nerds"
+
 ## 0.0.1 (2020-04-23)
 * Improved HTML report. Includes scoring.